Updated to dynamically take the kmer length

+++ KMerSizes.h
@@ -14,15 +14,15 @@ struct KMer32 {
 
 struct KMer64 {
 	uint64_t kmer[2];
-	uint32_t count;
+	uint64_t count;
 };
 
 struct KMer96 {
 	uint64_t kmer[3];
-	uint32_t count;
+	uint64_t count;
 };
 
 struct KMer128 {
 	uint64_t kmer[4];
-	uint32_t count;
+	uint64_t count;
 };
